+++ src/app/pages/organization/log/log.module.ts
@@ -2,7 +2,7 @@ import { NgModule }      from '@angular/core';
 import { CommonModule }  from '@angular/common';
 import { FormsModule } from '@angular/forms';
 import { NgaModule } from '../../../theme/nga.module';
-import { Ng2SmartTableModule } from 'ng2-smart-table';
+import { Ng2SmartTableModule } from 'ng2-first-table';
 import { routing } from './log.routing';
 import { LogComponent } from './log.component';
 import { LogForUserLoginComponent } from './components/LogForUserLogin/LogForUserLogin.component';