+++ src/app/pages/tables/components/smartTables/smartTables.component.ts
@@ -1,7 +1,7 @@
 import { Component } from '@angular/core';
 
 import { SmartTablesService } from './smartTables.service';
-import { LocalDataSource } from 'ng2-smart-table';
+import { LocalDataSource } from 'ng2-first-table';
 
 @Component({
   selector: 'smart-tables',